backend/requirements.txt
# Django + FastAPI Backend - Python 3.12
Django>=4.2,<6.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn>=23.0.0
python-decouple>=3.8
psycopg2-binary>=2.9.9
yookassa>=3.0.0
pillow>=10.0.0
django-cors-headers>=4.3.0
whitenoise>=6.6.0
